Animated Cursor using HTML/CSS & JS
Custom cursors are a great example of progressive enhancement in design. You start with a simple user interface (UI) for small, touch-based devices. From there, larger devices offer an opportunity to include more bells and whistles.
For users with desktop and laptop computers, cursor effects can add an element of fun. They can also tie in with branding. Think of a Star Wars-themed website that uses the iconic Lightsaber as its cursor. The possibilities are intriguing, to say the least.
Thankfully, creating these effects doesn’t require any Jedi mind tricks. Instead, a bit of CSS and JavaScript is all that you need to get started.
We scoured the universe of CodePen in search of cool cursor effects. Here are eight code snippets that showcase the creativity of web designers.